refactor: REST API에 맞게 게시판 기능 리팩토링

+++ src/main/java/com/goorm/nyangnyam_back/service/BoardService.java
@@ -2,67 +2,68 @@ package com.goorm.nyangnyam_back.service;
 
 import com.goorm.nyangnyam_back.model.Board;
 import com.goorm.nyangnyam_back.repository.BoardRepository;
-import org.springframework.beans.factory.annotation.Autowired;
+import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
-import org.springframework.web.multipart.MultipartFile;
 
-import java.io.File;
-import java.util.UUID;
 
 
 @Service
+@RequiredArgsConstructor
 public class BoardService {
 
-    @Autowired
-    private BoardRepository boardRepository;
+    private final BoardRepository boardRepository;
 
-    // 글 작성 처리
-    public void write(Board board, MultipartFile file) throws Exception{
-
-        String projectPath = System.getProperty("user.dir") + "\\src\\main\\resources\\static\\files";
-
-        UUID uuid = UUID.randomUUID();
-
-        String fileName = uuid + "_" + file.getOriginalFilename();
-
-        File saveFile = new File(projectPath, fileName);
-
-        file.transferTo(saveFile);
-
-        board.setFilename(fileName);
-        board.setFilepath("/files"+fileName);
-
-        boardRepository.save(board);
-
+    // 게시글 작성
+    public Board createBoards(Board board) {
+        return boardRepository.save(board);
     }
 
-    // 게시글 리스트 처리
-    public Page<Board> boardList(Pageable pageable){
+
+    // 모든 게시글 가져오기
+    public Page<Board> getAllBoards(Pageable pageable) {
         return boardRepository.findAll(pageable);
     }
 
-    public Page<Board> boardSearchList(String searchkeyword, Pageable pageable){
-        return boardRepository.findByTitleContaining(searchkeyword, pageable);
+
+    // 검색어 제목으로 게시글 가져오기 (게시글 검색)
+    public Page<Board> getSearchBoards(String search, Pageable pageable) {
+        return boardRepository.findByTitleContainingIgnoreCase(search, pageable);
     }
 
-    //특정 게시글 불러오기
-    public Board boardView(Integer id) {
+
+    // 게시글 상세
+    public Board getBoardsById(String id) {
         return boardRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("Invalid board Id:" + id));
     }
 
-    //특정 게시글 삭제
-    public void boardDelete(Integer id){
 
+    // 게시글 수정
+    public Board updateBoards(String id, Board board, String username){
+        Board existingBoard = boardRepository.findById(id)
+                .orElseThrow(() -> new IllegalArgumentException("Invalid board Id: " + id));
+        // 권한이 유효하지 않은 경우
+        if (!existingBoard.getUsername().equals(username)) {
+            throw new SecurityException("User is not authorized to update this board");
+        }
+
+        existingBoard.setTitle(board.getTitle());
+        existingBoard.setContent(board.getContent());
+        existingBoard.setPublicRange(board.getPublicRange());
+        existingBoard.setCategory(board.getCategory());
+        return boardRepository.save(existingBoard);
+    }
+
+
+    // 게시글 삭제
+    public void deleteBoards(String id, String username){
+        Board existingBoard = boardRepository.findById(id)
+                .orElseThrow(() -> new IllegalArgumentException("Invalid board Id: " + id));
+        // 권한이 유효하지 않은 경우
+        if (!existingBoard.getUsername().equals(username)) {
+            throw new SecurityException("User is not authorized to delete this board");
+        }
         boardRepository.deleteById(id);
     }
-
-    //좋아요 기능
-    public void increaseLikeCount(Integer id){
-        Board board = boardRepository.findById(id).orElseThrow();
-        board.setLikeCount(board.getLikeCount()+1);
-        boardRepository.save(board);
-    }
-
 }
